+++ ChangeLog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- **breaking change** `e3nn.normal` has a new argument to get normalized vectors.
 
 ## [0.10.1] - 2022-10-24
 ### Added